Add files via upload
The cosmological principle—the assumption that the universe is homogeneous and isotropic on large scales—underpins the standard model of cosmology. Although widely accepted, this principle has not been subjected to extensive testing using independent observational parameters beyond redshift and the cosmic microwave background (CMB). This paper introduces an alternative hypothesis based on galaxy formation ages as a potential method to verify or challenge this fundamental assumption.